- New design for Send Token screen

+++ app/src/main/java/com/p2p/wallet/swap/ui/SwapContract.kt
@@ -15,7 +15,7 @@ interface SwapContract {
         fun showDestinationToken(token: Token)
         fun showFullScreenLoading(isLoading: Boolean)
         fun showLoading(isLoading: Boolean)
-        fun showPrice(amount: Double, exchangeToken: String, perToken: String)
+        fun showPrice(amount: BigDecimal, exchangeToken: String, perToken: String)
         fun showCalculations(data: CalculationsData)
         fun setAvailableTextColor(@ColorRes availableColor: Int)
         fun showAroundValue(aroundValue: BigDecimal)
